fix(layers): stop asking for a new frame when nothing has changed
A still 3D map never went idle: it rendered continuously and pinned the CPU.
Attribution showed four redraw requests per frame, and in a standing-still view
they were the only ones.

CompositeVectorTileLayer::applyConfig runs from onDrawFrame for every raster and
hillshade child, every frame, and applied the "cheap redraw-only" setters -
opacity, exaggeration, the colours, method, contour width, illumination
direction - unconditionally. Every one of those ends in Layer::redraw(), so a
style defining four of them asked for four frames per frame, for ever. They are
now applied only when the value actually changed, guarded against the
last-applied map the decode-bound setters already used (widened to double, which
holds a 32-bit ARGB exactly). Zoom-interpolated values still animate, because
they genuinely change when the zoom moves.

ViewState::getTerrainMaxZoom had the same shape of bug in the camera clearance
clamp: zooming scales the camera-to-FOCUS vector, but the bound was derived from
the camera's height above z=0. On terrain the focus sits on the ground, so every
bound landed short of the clearance shell, the next frame clamped again, and the
map crept outwards a frame at a time. Measured from the focus, the bound lands on
the shell.

Layer::redraw forwards its caller's file and line (default arguments, so no call
site changes) - every layer setter funnels through it, and without that the
renderer's redraw tally names this one function for all twenty of them.

+++ all/native/layers/CompositeVectorTileLayer.h
@@ -172,7 +172,7 @@ namespace carto {
         std::vector<ExternalSource> _externalSources;
         std::vector<DrawItem> _drawItems;
         std::map<std::string, std::map<std::string, float> > _lastVectorConfig; // per-source applied contour params
-        std::map<std::string, std::map<std::string, float> > _lastChildConfig;  // per-source last-applied decode-affecting values (avoid per-frame re-decode)
+        std::map<std::string, std::map<std::string, double> > _lastChildConfig; // per-source last-applied config values (double: holds a 32-bit ARGB exactly)
         bool _singlePassRenderingEnabled;
 
         // Cached component handles for wiring child layers added after setComponents().
